+++ sentry-spring-boot-starter/src/test/kotlin/io/sentry/spring/boot/SentryAutoConfigurationTest.kt
@@ -154,7 +154,7 @@ class SentryAutoConfigurationTest {
         ).run {
             val options = it.getBean(SentryProperties::class.java)
             assertThat(options.readTimeoutMillis).isEqualTo(10)
-            assertThat(options.shutdownTimeout).isEqualTo(20)
+            assertThat(options.shutdownTimeoutMillis).isEqualTo(20)
             assertThat(options.flushTimeoutMillis).isEqualTo(30)
             assertThat(options.isDebug).isTrue()
             assertThat(options.diagnosticLevel).isEqualTo(SentryLevel.INFO)